# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Clipper API
  description: |
    Complete API documentation for the Clipper platform - a social platform for Twitch clip curation and sharing.
    
    ## Authentication
    Most endpoints require authentication via JWT Bearer token:
    ```
    Authorization: Bearer <your_jwt_token>
    ```
    
    ## Rate Limiting
    API endpoints are rate-limited to prevent abuse. Rate limits vary by endpoint and are documented in each operation.
    Common limits:
    - Public endpoints: 60-100 requests/minute
    - Authenticated actions: 10-30 requests/minute
    - Submission operations: 5-10 requests/hour
    
    ## Pagination
    List endpoints support pagination via `page` and `limit` query parameters.
    Default limit is 20, maximum is 100.
    
    ## Errors
    The API uses standard HTTP status codes:
    - 200: Success
    - 201: Created
    - 400: Bad Request
    - 401: Unauthorized
    - 403: Forbidden
    - 404: Not Found
    - 429: Too Many Requests
    - 500: Internal Server Error
    
  version: 1.0.0
  contact:
    name: Clipper API Support
    url: https://git.subcult.tv/subculture-collective/clpr
    email: support@clpr.tv
  license:
    name: MIT
    url: https://git.subcult.tv/subculture-collective/clpr/blob/main/LICENSE
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://staging.clpr.tv
    description: Staging environment
  - url: https://api.clpr.tv
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/favorites:
    get:
      tags: [Clips]
      summary: List user favorites
      description: Returns current user's favorite clips
      operationId: listUserFavorites
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: List of favorite clips
          content:
            application/json:
              schema:
                type: object
                properties:
                  favorites:
                    type: array
                    items:
                      $ref: '#/components/schemas/Clip'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ========================================
  # Tags
  # ========================================
...
  /api/v1/users/me/email-logs:
    get:
      tags: [Users]
      summary: Get email logs
      description: Returns email logs for current user
      operationId: getUserEmailLogs
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Email logs
          content:
            application/json:
              schema:
                type: object
        '401':
          $ref: '#/components/responses/Unauthorized'


# Component Schemas
# ========================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from authentication endpoints
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number for pagination
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of items per page
...
  responses:
...
    Unauthorized:
      description: Unauthorized - Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: object
          description: Additional error details
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 100
        total_pages:
          type: integer
          example: 5
        has_next:
          type: boolean
        has_prev:
          type: boolean
...
    Clip:
      type: object
      required:
        - id
        - twitch_clip_id
        - twitch_clip_url
        - title
        - creator_name
        - broadcaster_name
        - view_count
        - vote_score
        - created_at
      properties:
        id:
          type: string
          format: uuid
        twitch_clip_id:
          type: string
        twitch_clip_url:
          type: string
          format: uri
        embed_url:
          type: string
          format: uri
        title:
          type: string
        creator_name:
          type: string
        creator_id:
          type: [string, "null"]
        broadcaster_name:
          type: string
        broadcaster_id:
          type: [string, "null"]
        game_id:
          type: [string, "null"]
        game_name:
          type: [string, "null"]
        language:
          type: [string, "null"]
        thumbnail_url:
          type: [string, "null"]
          format: uri
        duration:
          type: [number, "null"]
          format: float
        view_count:
          type: integer
        vote_score:
          type: integer
        comment_count:
          type: integer
        favorite_count:
          type: integer
        is_featured:
          type: boolean
        is_nsfw:
          type: boolean
        is_removed:
          type: boolean
        is_hidden:
          type: boolean
        trending_score:
          type: number
          format: float
        hot_score:
          type: number
          format: float
        created_at:
          type: string
          format: date-time
        imported_at:
          type: string
          format: date-time
...